feat: added create api for posts.

+++ src/main/java/com/example/interviewtask/application/post/repository/PostRepository.java
@@ -4,6 +4,5 @@ import com.example.interviewtask.domain.post.Post;
 
 
 public interface PostRepository {
-
-    Long create(Post post);
+    int create(Post post);
 }
