+++ src/main/java/com/example/interviewtask/application/post/use_cases/CreatePostUseCase.java
@@ -3,5 +3,5 @@ package com.example.interviewtask.application.post.use_cases;
 import com.example.interviewtask.application.post.dto.CreatePostDto;
 
 public interface CreatePostUseCase {
-    long execute(long authorId, CreatePostDto dto);
+    int execute(int authorId, CreatePostDto dto);
 }
